version 0.9.0 complete
improve the initial function logic

+++ app/scripts/detail.js
@@ -15,7 +15,6 @@
         ],
         function (ec) {
 
-
             // DOM Elements
             var mainViewer = document.getElementById('main-viewer');
 
@@ -84,23 +83,16 @@
                         trigger: 'axis'
                     },
                     legend: {
-                        data: ['导航流量', '产品流量']
+                        data: ['avg product', 'avg category']
                     },
                     toolbox: {
-                        show: true,
-                        feature: {
-                            mark: {show: true},
-                            dataView: {show: true, readOnly: false},
-                            magicType: {show: true, type: ['line', 'bar']},
-                            restore: {show: true},
-                            saveAsImage: {show: true}
-                        }
+                        show: false
                     },
                     calculable: true,
                     xAxis: [
                         {
-                            type: 'category',
-                            data: ['1月', '2月', '3月', '4月', '5月', '6月', '7月', '8月', '9月', '10月', '11月', '12月']
+                            type: 'category'
+                            //data: ['China', 'Japan', 'USA', 'France']
                         }
                     ],
                     yAxis: [
@@ -110,96 +102,109 @@
                     ],
                     series: [
                         {
-                            name: '导航流量',
+                            name: 'avg product',
                             type: 'bar',
-                            data: [2.0, 4.9, 7.0, 23.2, 25.6, 76.7, 135.6, 162.2, 32.6, 20.0, 6.4, 3.3],
-                            markPoint: {
-                                data: [
-                                    {type: 'max', name: '最大值'},
-                                    {type: 'min', name: '最小值'}
-                                ]
-                            },
-                            markLine: {
-                                data: [
-                                    {type: 'average', name: '平均值'}
-                                ]
-                            }
+                            data: [500, 400, 600, 400]
                         },
                         {
-                            name: '产品流量',
+                            name: 'avg category',
                             type: 'bar',
-                            data: [2.6, 5.9, 9.0, 26.4, 28.7, 70.7, 175.6, 182.2, 48.7, 18.8, 6.0, 2.3],
-                            markPoint: {
-                                data: [
-                                    {name: '年最高', value: 182.2, xAxis: 7, yAxis: 183, symbolSize: 18},
-                                    {name: '年最低', value: 2.3, xAxis: 11, yAxis: 3}
-                                ]
-                            },
-                            markLine: {
-                                data: [
-                                    {type: 'average', name: '平均值'}
-                                ]
-                            }
+                            data: [600, 800, 200, 600]
                         }
                     ]
                 }
             };
 
+            var operationHistory = [];
 
             // Application logic
             function appReady() {
 
-                //todo: init the analysis demo
-                getAnalysisResult();
+                // 1. get the init model
+                // 2. get the further analysis target
+                // 3. refresh the current model
+                setInitModel();
+                // TODO: 3. refresh the current model & add the latest model to the history list
 
             }
 
             // Application logic detail
-            function getAnalysisResult() {
-
+            function setInitModel() {
                 var vars = getUrlParams();
-                var type = vars['model'],
+                var model = vars['model'],
                     paramX = vars['x'],
                     paramY = vars['y'];
-                var returnData = [];
+                var returnData = {
+                    legendData: [paramX, paramY],
+                    seriesData: []
+                };
 
-                $.getJSON('../data/demo-model.json', function (json) {
-                    returnData = json['result'];
+                operationHistory = [];
+                operationHistory.push(model);
 
-                    if (type !== undefined) {
-                        // TODO: analysis the model and its parameters
+                // TODO: query from database (x, y)
 
-                        if (type === 'pie') {
-                            ecGraph.pieOption.legend.data = [paramX, paramY];
-                            ecGraph.pieOption.series[0].data = returnData;
-                            ecGraph.mainViewer.setOption(ecGraph.pieOption, {notMerger: true});
-                            ecGraph.mainViewer.on(ecGraph.ecConfig.EVENT.CLICK, mainViewerClickHandler);
-                        } else if (type === 'bar') {
-                            ecGraph.mainViewer.setOption(ecGraph.barOption, {notMerger: true});
-                        } else if (type === 'heatmap') {
-
-                        } else {
-
-                        }
+                $.getJSON('../data/demo-model.json',function (json) {
+                    returnData.seriesData = json['result'];
+                    if (model !== undefined) {
+                        setGraphModel(model, returnData);
                     } else {
-                        alert('Please choose a model!');
+                        console.log("No model!!!");
+                    }
+                })
+            }
+
+            function refreshMainViewer() {
+                var furtherModel = localStorage.getItem('furtherAnalysisModel'),
+                    furtherData  = localStorage.getItem('furtherData'),
+                    refresh      = localStorage.getItem('refresh'),  //typeof(refresh) === string
+                    returnData   = {
+                        legendData: ['avg product', 'avg category'],
+                        seriesData: []
+                    };
+
+                operationHistory.push(furtherModel);
+
+                // TODO: query from database (furtherData)
+
+                $.getJSON('../data/further-model.json', function (json) {
+                    returnData = json;
+
+                    if(refresh === 'false'){
+                        setGraphModel(furtherModel, returnData['countries']);
+                    } else {
+                        setGraphModel(furtherModel, returnData['region']);
                     }
                 });
 
+                // TODO: refresh the history panel
+                setHistoryPanel();
 
             }
 
-            function refreshAnalysisResult(type) {
-                if (type === 'pie') {
+            function setGraphModel(model, returnData) {
+                if (model === 'pie') {
+                    ecGraph.pieOption.legend.data = returnData.legendData;
+                    ecGraph.pieOption.series[0].data = returnData.seriesData;
                     ecGraph.mainViewer.setOption(ecGraph.pieOption, {notMerger: true});
-                } else if (type === 'bar') {
-                    ecGraph.mainViewer.setOption(ecGraph.barOption, {notMerger: true});
-                } else if (type === 'force') {
+                    ecGraph.mainViewer.on(ecGraph.ecConfig.EVENT.CLICK, mainViewerClickHandler);
+                } else if (model ==='bar') {
 
+                    console.log(returnData.names);
+                    ecGraph.barOption.xAxis[0].data = returnData.names;
+                    ecGraph.barOption.series[0].data = returnData['avgProduct'];
+                    ecGraph.barOption.series[1].data = returnData['avgCategory'];
+                    ecGraph.mainViewer.setOption(ecGraph.barOption, {notMerger: true});
                 } else {
 
                 }
-                ecGraph.mainViewer.on(ecGraph.ecConfig.EVENT.CLICK, mainViewerClickHandler);
+            }
+
+            function setHistoryPanel() {
+                $('#analysis-history').empty();
+                for(var i=0; i<operationHistory.length; i++){
+                    $('#analysis-history').append('<img src="../images/'+operationHistory[i]+'.png" class="img-rounded" height="90">');
+                }
             }
 
             // Event Binding
@@ -213,29 +218,46 @@
 
             function mainViewerClickHandler(param) {
                 var data = param.data;
-                console.log("You click the " + data.name);
-                localStorage.setItem("furtherData", data.name);
 
-                $.getJSON('../data/model-graph.json', function (json) {
-
-                    var result = json[data.name]['models'];
-
-                    for (var i = 0; i < result.length; i++) {
-                        var modelItem = $('<li class="media image-menu-item"><a class="media-left" ' +
-                        'href="detail.html?model=' + result[i] + '&x=' + data.name + '" name="' + result[i] + '">' +
-                        '<img src="../images/' + result[i] + '.png" width="70"></a><div class="media-body">' +
-                        result[i] + ' Graph</div></li>');
-                        $("#further-model-list").append(modelItem);
+                if(data.name === undefined) {
+                    if (confirm("Refresh the data from another perspective?")) {
+                        localStorage.setItem('refresh', true);
+                        refreshMainViewer();
                     }
+                } else {
+                    localStorage.setItem('refresh', false);
+                    console.log("You click the " + data.name);
+                    localStorage.setItem("furtherData", data.name);
+                    $("#further-model-list").empty();
+                    $.getJSON('../data/model-graph.json', function (json) {
 
-                });
+                        var result = json[data.name]['models'];
 
-                furtherTitle.html("Further analysis for <strong>" + data.name + "</strong>");
+                        for (var i = 0; i < result.length; i++) {
+                            $modelItem = $('<li class="media image-menu-item" name="'+result[i]+'"><a class="media-left" ' +
+                            'href="#"><img src="../images/' + result[i] + '.png" width="70"></a><div class="media-body">' +
+                            result[i] + ' Graph</div></li>');
+                            $modelItem.click(furtherAnalysisHandler);
+                            $("#further-model-list").append($modelItem);
+                        }
+
+                    });
+
+                    furtherTitle.html("Further analysis for <strong>" + data.name + "</strong>");
+                }
+            }
+
+            function furtherAnalysisHandler() {
+                //var furtherModel = $(this).attr('name');
+                localStorage.setItem('furtherAnalysisModel', $(this).attr('name'));
+
+                // click the see the further analysis
+                refreshMainViewer();
             }
 
             function modelSelectHandler(param) {
                 var data = $(this).attr('name');
-                refreshAnalysisResult(data);
+                // TODO: change current analysis view
             }
 
             function faOkHandler() {
@@ -243,10 +265,8 @@
                 console.log(faFromContent);
             }
 
-
             // Execution logic
             appReady(); // application startup
-
 
             // Util functions
             function getUrlParams() {
